Documentation

All docstrings!

SPICEKernels.SPICEKernels — Module

Download any and all General Kernels provided by NASA JPL!

Warning

This package is not affiliated with or endorsed by NASA, JPL, or any other organization! This is an independently written package by an astrodynamics hobbyist.

Extended Help

README

Imports

AstroTime

Base

Core

Dates

DocStringExtensions

HTTP

Scratch

Exports

SPICEKernels.DigitalShapeKernel — Type

abstract type DigitalShapeKernel <: SPICEKernels.SPICEKernel

A supertype representing Digital Shape Kernels (DSK).

SPICEKernels.EphemerisKernel — Type

abstract type EphemerisKernel <: SPICEKernels.SPICEKernel

A supertype representing Spacecraft and Planet Kernels, also known as ephemeris kernels (SPK).

SPICEKernels.FramesKernel — Type

abstract type FramesKernel <: SPICEKernels.SPICEKernel

A supertype representing Frames Kernels (FK).

SPICEKernels.KernelType — Type

primitive type KernelType <: Enum{Int32} 32

An enumerated type with all possible SPICE kernel flavors.

SPICEKernels.LeapsecondsKernel — Type

abstract type LeapsecondsKernel <: SPICEKernels.SPICEKernel

A supertype representing Leapseconds Kernels (LSK).

SPICEKernels.PlanetaryConstantsKernel — Type

abstract type PlanetaryConstantsKernel <: SPICEKernels.SPICEKernel

A supertype representing Planetary Constants Kernels (PCK).

SPICEKernels.SPICEKernel — Type

abstract type SPICEKernel

A supertype for all SPICE Kernels! To satisfy the SPICEKernel interface, your new type must implement each of the methods below.

SPICEKernels.bodies(kernel)::Union{Nothing, <:AbstractSet{<:Integer}}

SPICEKernels.type(kernel)::SPICEKernels.KernelType

SPICEKernels.name(kernel)::AbstractString

SPICEKernels.SPK — Type

struct SPK <: SPICEKernels.EphemerisKernel

A spacecraft and planet kernel (SPK).

SPICEKernels.bodies — Function

Return the bodies contained within the provided kernel, if they are explicitly provided.

SPICEKernels.getkernel — Method

getkernel(identifier, type; to, cache)

Given the ephemeris file name, download the file and return a path to the file location. If a full URL is provided, that URL will be used. Otherwise, the identifier is assumed to be a SPICE General Kernel. If to is set, the file is downloaded to the provided path. If cache is enabled –- and it is by default –- the ephemeris file is downloaded to the SPICEKernels.jl scratch space.

SPICEKernels.name — Function

Return the file path to kernel.

SPICEKernels.type — Function

Return the underlying SPICE kernel type.